feat: use version from top-level lerna.json/package.json
If there is a lerna.json file, we are going to use that version when
publishing the builds. If there isn't, we use the version from the
top-level package.json.

This means that all sub-packages will get the same version as the
parent, linking all versions of the main package together.